Changing unnecesary includes
Still does not solve that, tried using namespace still not working.

ALSO, when undefining the use of a namespace, every class that requires the Polygon.h class breaks completely as there is a problem with its definition I can't quite understand nor find.

USING NAMESPACE MATH everywhere does not solve anything as it already finds the functions in compiler but not in linker.