Some FBX may draw more or less okey
The warror fbx draws more or less, with a couple of errors, but any other fbx thrown is just a broken mess.

I don't understand how to draw with Elements as the mesh returns as many vertices as indices (with a certain relation) instead of just all 'real' vertices and then the relation with indices, as it is done when we create a cube for example

As for mem leak that happende, we were generating the buffer each time instead of generating in on creation then later just binding it.