I have no idea
i don' tknow why it is not rendering the texture.

Can't say I am doing anything correctly as I don't know.
The tex_coords are supposedly generated correctly, then the buffers are set and generated

The Checker texture should be generated correctly too, and when rendering trying to set different values only change the brightness? It does not actually set a checker pattern~

Proper Enable of GL_TEXTURE_2D, and proper binding of the Texture and the TexCoords too I think.