Render Machine Broke
Trying to do the render I broke it.
basically the new system of loading textures is not working (not even the checker because some reason I do not understand right now)

There is a clear issue with reading the data and setting up the textures as the fbx appears to have none (even when the house should have one already) thus breaking the load.

Still the checkered texture should be working.