Previous commit bug solved + Cleaning
The BBs of objects without meshes was not updating as they should.

Now the events for transformations are called in ModuleGameObject->ManageGuizmo and in the InspectorWindow. (before it was done in componentTransform, but when initializing a lot of duplicated events were sent).

Cleaned fcns that are not needed from gamobject.cpp.